azurestack: Fix destroyer to delete A records
Azure Stack Hub destroyer checks for the environment name before
deleting the A records but it checks the environment name after
the Azure SDK returns an environment and the sdk right now is
returning "HybridEnvironment" instead of "AzureStackCloud" that
the destroyer is expecting.

Adding a fix to pick our cloud name instead on the one returned
by the Azure sdk.