Have the MCO manage the worker ignition stub config on baremetal
On the baremetal platform, worker MachineSets on newly-installed
clusters will automatically use the latest version of CoreOS to
provision, as described in
https://github.com/openshift/enhancements/blob/master/enhancements/baremetal/coreos-image-in-release.md

To ensure that the MachineConfig Ignition stub always uses a compatible
version of the Ignition format, point these MachineSets at the Secret
managed by the MCO (worker-user-data-managed), which is created since
https://github.com/openshift/machine-config-operator/pull/2827. Since a
non-managed version of the Secret is not required on this platform,
install the worker-user-data-managed Secret initially rather than the
previous worker-user-data Secret.

Other platforms will continue to install worker-user-data Secrets,
which are preserved as they were at cluster creation time, because they
are referenced by MachineSets that have fixed CoreOS images. When these
MachineSets are updated to always use the latest CoreOS images, they
should also be updated to point to the managed user data.

This is the part of the installer changes for the enhancement:
https://github.com/openshift/enhancements/blob/master/enhancements/machine-config/user-data-secret-managed.md

A previous version of this patch (but for all platforms) was previously
committed as 8d278d258f9eb16788cc8c4bda842018e28625d7, but later
reverted by 3920ae4a6bbe0690b1b23c603220257abe69a664.