azurestack: Fix gather bootstrap for azurestack
The terraform stages needs just a list of control plane ips to
gather bootstrap logs but the azurestack terraform output is
a list of lists causing the gather command to fail. Fixing the
output of the terraform stage to flatten the ips to a single
list. This will enable azurestack gather command.