This commit aligns the installer with the current Alibaba provider library.
The updates include renaming and replacing the ResourceTagReference as well as
modifying the SystemDisk property. The tag update changes the type
to an AlibabaResourceReference which is a struct that contains an ID
field and a slice of tags. This reduces the need for two separate API fields and simplifies the
logic into a single field. The reference can hold the information in
which to search for a component inside of the provider or reference it
by ID.

The SystemDisk info update is an attempt to reduce the surface area of the
API and consolidate all information of the SystemDisk into a single
struct.