Add a nil check after Delete
An error path was invoking a function from an embedded interface in the
return value from a Delete() call. It turns out there are error
conditions where that embedded interface has not yet been initialized,
leading to a nil pointer exception. This commit adds a nil check for the
embedded member. For readability, we also explicitly dereference the
function via the interface name.

Closes: 2026853