Azure Stack: add CA support in bootstrap ign stub
On Azure Stack, this adds CAs from the additionalTrustBundle to the
ignition stub.

Azure Stack uses shared-access signatures (SAS) to
manage access to the ignition storage account. Creating a SAS
requires an existing storage account, so we cannot use a similar
approach to AWS, which creates a pre-signed URL and passes it
through the tfvars.

Instead, this approach manipulates the ignition stub in the Terraform
config to append the SAS to the storage account URL.