azure: Fix terraform cluster IP output
Fixing an output produced by the cluster stage of create command
in Azure where the stage provided a list of lists as output for
the IP addresses. This could be made as simple as a list hence
making the change to pick only the first IP address of each
control plane node.