Terraform: Build providers individually
This will build all the Terraform providers individually. Each provider has
it's own directory along with a <provider-name>.go file that imports the
provider so that dependencies can be determined. There is probably a better way
to do this, but it works.

The Makefile greps the import statement to determine the path to build each
binary.

Our custom providers (azureprivatedns and vsphereprivate) are built from the
directory and not a vendor directory.

Building these has been added hack/build.sh. In the future a better way of
handling the vendoring in the individual directories will be necessary.