terraform: allow for building local providers individually
This will build local terraform providers individually. The local terraform
providers are the providers that either (1) are not published in a public
regisrty or (2) require patches.

azureprivatedns: private to the installer repo
azurestack: requires patches
ibm: arm build is not published
ignition: arm buid is not published
ironic: not published
ovirt: not published
vsphere: requires patches
vsphereprivate: private to the installer repo

Building these has been added hack/build.sh. In the future a better way of
handling the vendoring in the individual directories will be necessary.

co-authored: staebler