Terraform: Use terraform-exec and terraform-json libraries
Terraform Init, Create and Destroy functions that used internal Terraform
functions have been replaced with equivalent functions provided by the
terraform-exec library.

Handling of terraform state has been replaced and rewritten to use the
terraform-json library.

Terraform providers now are built individually, tarred and bzipped into a file
placed into the terraform directory where it is embedded into the installer
binary. The plugins init function decompresses and extracts this into a user
configurable directory and fills in the plugin map so the paths for terraform
and each provider are known.