ibmcloud: remove resource group question from survey
The installer supports creating a resource group for the user when
installing to ibmcloud. This should be the default, opinionated
way that the installer in which the installer is configured. As such,
the installer should not be asking in the survey for the user to
provide an existing resource group to use.