aws: get ips for all control plane instances for bootstrap gather
The tf code for aws outputs only the IP address of the first control
plane instance rather than the IPs of all of the control plane instances.
The issue is that all of the IPs for all of the instances are collected
into a single list and only then the first IP is taken from the list.
Instead, we want to take the first IP from each instance and then collect
those IPs into a list.