Set rc-manager=unmanaged for on-prem bootstrap
There is a dispatcher script on the bootstrap node that is supposed
to be responsible for updating resolv.conf on network changes. We
don't want NetworkManager to step on those changes, so it needs to
be set to unmanaged. This is already the case on cluster nodes[0],
but apparently was missed in the bootstrap config.

0: https://github.com/openshift/machine-config-operator/blob/4b98289781ec5f3fe7313049496afda654f6e0fc/templates/common/on-prem/files/NetworkManager-onprem.conf.yaml#L7