azure: Remove allowed values for nodeVMsize
Providing allowedValues for the node VMsize for an Azure UPI installation
seems to be an overhead for the installer where we have to
constantly update the list of values that can be used every
time azure changes or creates a new type.

Removing the allowed value list to prevent this overhead.