Set 1 ingresscontroller replica for private SNO
When generating a default ingresscontroller manifest, check the
infrastructure topology, and set the ingresscontroller's replicas to 1 if
the topology is "SingleReplica" (i.e., Single-Node OpenShift).

When the publishing strategy is "Internal", the installer generates a
default ingresscontroller manifest.  Before this commit, the installer did
not check the cluster topology and always left replicas unspecified in the
generated ingresscontroller.  As a result, the ingress operator used the
default value of 2 replicas, causing the ingress operator to try to
schedule 2 pod replicas for the ingresscontroller, which would fail on
single-node clusters because only 1 pod replica can be scheduled per
ingresscontroller per node.

This issue does not arise when the publishing strategy is "External"
because in this case, the installer does not generate a default
ingresscontroller manifest; instead, the ingress operator creates the
default ingresscontroller, and the ingress operator does check the topology
and set replicas appropriately in the ingresscontroller that it creates.

This commit fixes bug 2042826.

https://bugzilla.redhat.com/show_bug.cgi?id=2042826

* pkg/asset/manifests/infrastructure.go (Generate): Factor out logic for
computing the control-plane topology and infrastructure topology.
* pkg/asset/manifests/ingress.go (generateDefaultIngressController): Check
the infrastructure topology and set replicas to 1 if the topology is
"SingleReplica".
* pkg/asset/manifests/ingress_test.go: New file.
(TestGenerateDefaultIngressController): New test.  Verify that
generateDefaultIngressController behaves correctly.
(withPublish): New helper for icBuildNamespace.
(ingresscontrollerOption, ingresscontrollerBuildNamespace): New types for
building test inputs.
(ingresscontrollerBuild): New variable for building test inputs.
(build, withLoadBalancer, withReplicas, withScope): New helpers for
ingresscontrollerBuildNamespace.
* pkg/asset/manifests/utils.go (getControlPlaneTopology)
(getInfrastructureTopology): New functions.
* pkg/asset/manifests/utils_test.go: New file.
(TestGetInfrastructureTopology): New test.  Verify that
getInfrastructureTopology behaves correctly.
(withControlPlaneReplicas, withWorkerReplicas): New helpers for
TestGetInfrastructureTopology.