update alibaba for provider spec api changes
This change updates the alibaba provider spec usage related to the
vswitch, security groups, and resource group. The API for the provider
spec is changing to use a discriminated union to capture the various
methods for finding resources (by id, name, or tags).

It also updates several machine api references to note the bifurcated
nature of the api version between v1beta1 and v1.