aws: Remove non-public AWS regions from list of regions
When creating the install-config, the installer displays regions
of all partitions of AWS. Certain regions also need extra information
for the validation to work and should not be taken as input since we
only ask for the bare minimum amount of information to create the
install config.

The best approach here would be to only display all the public regions
of AWS and allow for other regions after the install-config is created
to allow for the user to add the extra information.