terraform: allow for downloading remote providers individually
This will download remote terraform providers individually. The remote terraform
providers are the providers that can be used as-is from a public registry.

This will also grab the terraform binary to embed in the installer from the
same terraform image used to download the terraform providers.

Use the latest version available for all of the remote providers.