terraform: pass var file paths to stage destroy
The normal stage destroy performs a terraform destroy, which
uses DestroyOptions. The gcp custom stage destroy for the
post-bootstrap destroy performs a terraform apply, which requires
ApplyOptions. All of the options passed to the destroy are VarFileOptions.
So instead of passing them, pass the paths to the var files and let
the destroy functions create either the DestroyOptions or ApplyOptions,
as needed.