Add IBM Power VS: tf plugin for ibm presign
This plugin is needed in order to securely host the ignition config
files for bootstrap machines. It uses the COS offering from IBM Cloud,
which is s3-compatible.

For more background on IPI on Power VS, refer to the enhancement
proposal here: openshift/enhancements#736

Older discussions on some of the code here can be found in #5224

Signed-off-by: Christy Norman <christy@linux.vnet.ibm.com>