Run assisted-service only on pre-identified node0
Run the assisted service only on node0.

Run `make iso` which creates an ISO with a default node zero ip `192.168.122.2`.

To set a different IP for node zero, `go run cmd/main.go -node-zero-ip 192.168.122.4`

A systemd service  `node-zero.service` looks for the node zero IP and if the current node's IP matches with it then only the `assisted-service.service` systemd service is startred. The `assisted-service.service` is responsible for running the assisted service.