vsphere: Validate num cores with vcpus
An error occurs if the vcpus specified is not a multiple of the
number of cores specified. A validation check is added to check
if it is indeed a multiple. The check also accounts for default
values of both the vcpus and number of cores (4 by default) in
case only one of the two is set.