Azure: use azidentity with adapter
Microsoft is in the process of upgrading the Azure SDK for Go from V1 to
V2. Our clients and authentication are on V1. The V1 authentication
utilizes ADAL which will be deprecated June 30, 2022. All V2 clients,
except the V2 auth client azidentity, are in beta (azidentity is
scheduled to be stable in Q2 2022). [0]

These changes remove the dependency on the ADAL API, replace the
authentication with azidentity, and use an adapter so the auth will work
with V1 clients.

[0] https://azure.github.io/azure-sdk/releases/latest/index.html#go