AGENT-37 Create cluster and infra-env using assisted-service REST-API
The ZTP manifests are placed under data/manifests.

The tool reads in the manifests and creates parameters in JSON
that are used to specify the new cluster and infra-env.

A new service, create-cluster-and-infra-env, waits for
assisted-service to be running and for the REST-API
to be ready. The service then POSTs the JSON parameters to
the REST-API, creating a cluster and a infra-env. Having
a dependency to assisted-service ensures that the other
hosts running the ISO doesn't create duplicate clusters
and infra-envs.