terraform: upgrade alicloud provider to v1.158.0
When #5507 was merged, the version of the alicloud terraform provider was set to v1.148.0,
which is the version in use when the PR was created. However, in the time between creation
and merge, #5571 was merged, which upgrade the version to v1.153.0. This change was lost
during conflict resolution.

Let's upgrade to the latest version (v1.158.0) of the alicloud terraform provider while we
are at it.