Update SR-IOV & DPDK doc for OpenStack
In 4.11, we did the following:

* Remove the dependency on config-drive and support OpenStack Metadata
  URL fetching.
* Automate the configuration of VFIO NO IOMMU
* Add support for vhostuser interfaces for DPDK

This patch adds more content for DPDK, with an example of a pod using
hugepages, dedicated CPUs and the DPDK port.

It also add more content for SR-IOV, like how to create a policy & a
pod.

Later, we'll add content for SR-IOV with HW offload.