Set single-node none-platform clusters Ingress `DefaultPlacement` to `ControlPlane`
Partial implementation for https://github.com/openshift/enhancements/blob/master/enhancements/single-node/single-node-openshift-with-workers.md
(the other missing part being generating a worker.ign even for
bootstrap-in-place installs)

The infrastructure asset is now a dependency of the ingress asset,
because the topologies determined within it affect how the new
`DefaultPlacement` Ingress config status field is set.

A none-platform single node cluster doesn't need a load balancer,
the API and ingress traffic for such cluster can be directed at the
single node directly. We want to maintain that even when worker
nodes are added to such cluster. We do that by asking the Cluster
Ingress Operator to place the ingress pod on the single control
plane node. This would ensure that the ingress pod won't ever be
scheduled on the added workers, as that would create a requirement
for a load-balancer.

Even when a single control-plane node cluster is installed with one or
more worker nodes since day 1, we still want control-plane ingress
default placement, for the sake of consistency. Users can override this
decision manually if they wish.