Set BOOTSTRAP_HOST_MAC in assisted-service environment
Setting the BOOTSTRAP_HOST_MAC environment variable notifies
assisted-service of a MAC address of Node Zero so that it will be set
as the bootstrap host.

We choose the MAC address of interface that the service IP is on, as
that is the most likely to be unique.

To pass the environment variable to assisted-service, we turn the node0
file used to trigger starting the service into an environment file.