azure: support selecting marketplace image in install config
Additional fields added to Azure machine pools that give the user
the option to select the marketplace image to use for the OS image.
The OS image can only be selected for compute machine pools.
Validation will be performed that (1) the marketplace image selected
exists and (2) that the license terms for the marketplcae image
selected have been accepted.

```
osImage:
  publisher: <string>
  offer: <string>
  sku: <string>
  version: <string>
```

https://issues.redhat.com/browse/CORS-1824