Make detecting Node0 more robust
If the Node0 IP appears on *any* interface, this is node 0.

Previously we tried to guess the 'best' IP for the host and only
designated it as node 0 if it was also the Node0 IP. Also, we were
comparing a URL host (IPv6 addresses wrapped in []) with a raw IP
address.