for vsphere ipi add cluster domain to the uploaded vm configs so that 30-local-dns-prepender can use it
Due to changes in FCOS 35 /etc/nsswitch.conf, we need to move the name resolution of api-int.<domain> to
systemd-resolved.  The way we do this is by adding a Domain attribute to /etc/systemd/resolved.conf.d/60-kni.conf.

Since we are using a slightly different prepender than the masters and workers the domain is not being
discovered automatically.  To address this we are adding the domain in data/data/vsphere/bootstrap/main.tf
which gets uploaded to the bootstrap instance on creation.

30-local-dns-prepender then downloads this information and adds it to /etc/systemd/resolved.conf.d/60-kni.conf

Note that we check for a non-empty domain before writing /etc/systemd/resolved.conf.d/60-kni.conf.

Fixes https://github.com/openshift/okd/issues/1182

Signed-off-by: John Fortin <fortinj66@gmail.com>