data/bootstrap/files/usr/local/bin/bootkube: Drop config CRDs
We've had config-operator rendering on the bootstrap node since
9994d371df (bootkube: render config.openshift.io resources,
2019-02-12, #1187).  Motivation for that commit isn't clear to me; [1]
suggests maybe keeping CRDs out of the installer repository.  But we
run a rendered cluster-version operator on the bootstrap machine since
63e27509dd (ignition: add CVO render to bootkube.sh, 2018-09-27, #330),
so we should be able to push resources at bootstrap time via the CVO.
Remove CRDs from the config rendering, so we can see if things work
without the config-rendered cluster-bootstrap pushes racing the
bootstrap CVO pushes, or the config-rendered pushes not realizing they
should filter out manifests annotated for capabilities that are not
enabled.

[1]: https://github.com/openshift/installer/pull/1187#issuecomment-463634025