vsphere: no longer logout of platform client before finished
Recent changes were applied that enabled the vsphere client to logout
when it was finished. However, in the case of the platform client usage,
it was logging out after validating the credentials but before it was
finished using the client. This was causing an issue when attempting to
connect to the cluster to prodvide additional prompt values.

This change moves the logout to the end of the platform usage so that it
only logs out when the platform is finished using the client.