Fix AWS Platform Provisioning Check
This fixes a bug in the AWS Platform Provisioning Check where it
identifies a record that merely matches the DNS suffix domain without
checking for a "." in the record.

For example, if cluster name is "abcd" with baseDomain "example.com"
then a record "foo-abcd.example.com" is misidentified as part of the
domain of the cluster ("abcd.example.com").