Modified PrintResource to now display more information
PrintResorces was changed so that it would have the ability to print out the valid values, defaults, and format when using explain for sub-fields in a struct. Without this change those fields would only be displayed when explaining the entire struct. Previously, using explain on a field of the struct would only print the description.

Updated tests for explain to accommodate changes to PrintResource