azure: move zones functionality to the client API
The zones.go file predates the existance of client.go. Most of the
functionality for retrieving resources through the Azure SDK is now
concentrated on the client file, except for Availability Zones. So let's
merge the zones.go code into client.go.

As an added benefit, we now have the ability to mock calls to fetch
the Availability Zones in the tests if needed.