Delete all the ports from tagged Neutron networks.
Currently, we are trying to remove ports, which are tagged, but
sometimes it happens, that port is created but tagging has failed, or we
didn't get response from Neutron API, resulting with untagged port.

Having network already tagged we can use its id to select all the port
within such network, so that all the ports, not only tagged would be
deleted.