Fixup Azure marketplace pr 5799 (#5839)
* vendor: azure client for marketplace images

Vendor in additional packages from github.com/Azure/azure-sdk-for-go
to support the calls needed to validate Azure marketplace images.

* azure: support selecting marketplace image in install config

Additional fields added to Azure machine pools that give the user
the option to select the marketplace image to use for the OS image.
The OS image can only be selected for compute machine pools.
Validation will be performed that (1) the marketplace image selected
exists and (2) that the license terms for the marketplcae image
selected have been accepted.

```
osImage:
  publisher: <string>
  offer: <string>
  sku: <string>
  version: <string>
```

https://issues.redhat.com/browse/CORS-1824

* azure: update client mock

Re-generate the mock for the Azure client so that the additional
functions added to support marketplace images are included in the
mock.

* azure: unit tests for static validation of marketplace images

Add unit tests for the static validation of marketplace images
selected in the install config.

* azure: unit tests for dyanmic validation of marketplace images

Add unit tests for the dynamic validation of the Azure marketplace images.

* fixup! azure: unit tests for dyanmic validation of marketplace images

Co-authored-by: staebler <staebler@redhat.com>