azure: client: use a recent profile for fetching VMs
We were using an azure-sdk profile for fetching VirtualMachines that was
compatible with Azure Stack. The shortcoming of such approach is that we
do not have access to information available to the instances in Public
clouds, like `LocationInfo`.

Therefore I'm proposing we use a more recent profile, not necessarily
compatible with Azure Stack. Other than offering extra information for
the VMs, this should not affect the current functionality for Public
clouds. It could however affect Stack cloud. In my tests that did not
seem to be the case though. The `LocationInfo` field is returned and
only its `ZoneDetails` is set to `nil`:

```
(*[]compute.ResourceSkuLocationInfo)(0xc000177e48)((len=1 cap=4) {
 (compute.ResourceSkuLocationInfo) {
  Location: (*string)(0xc0002065f0)((len=4) "ppe3"),
  Zones: (*[]string)(0xc000177e78)({
  }),
  ZoneDetails: (*[]compute.ResourceSkuZoneDetails)(<nil>),
  ExtendedLocations: (*[]string)(<nil>),
  Type: (compute.ExtendedLocationType) ""
 }
})
```

An alternative approach would be to create a new function in the client
to return the `LocationInfo`. I chose not to do that because it would
result in 2 API calls to the same resource for cases in which you need
to first validate the instance type and then retrieve the location info.