Agent-160: Add assets for create image command
Download the baseIso used for the agent "create image" command.

Note that cache.go is copied from ./pkg/tfvars/internal/cache/,
since its in an internal dir it can't be imported. A follow-on PR
will move this file to a common area so that it can be used by
both pkgs.