AGENT-131: Ignition and Iso assets for agent installer
The Ignition asset embeds the asset/ignition/bootstrap/Common type.
The Ignition assets fills in the SSHPublicKey and the manifests that
are currently available. Much of the ignition data will be populated
in a future PR.

The ISO asset writes to agent-installer.iso leaving the fleeting.iso
unchanged. The ISO asset will replace fleeting.iso when they are
equivalent.