azure: partially reverts some changes to validation
Instead of getting the capabilities as a map and checking for individual
values, revert to the previous behaviour of looping through an instance
type's capabilities and validating the values present.

This guarantees that we do not fail in case the value of some capability
could not be determined early on.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2084441