Bug 2085380: azure: validate VM image and instance HyperV gen match
Otherwise when inspecting why a node failed to deploy, a user will
encounter

"The selected VM size '<size>' cannot boot Hypervisor Generation '1'. If
this was a Create operation please check that the Hypervisor Generation
of the Image matches the Hypervisor Generation of the selected VM Size.
If this was an Update operation please select a Hypervisor Generation
'1' VM Size."

when using an instance with HyperVGeneration == V2 and marketplace image
V1.