Bug 2064693: Restore ability to use local clouds.yaml
We recently switched from using Terraform as a library to using it as a
binary. As part of this switch, we have changed the definition of "$PWD"
for the providers - it is now a temporary working directory we create
rather than the user's current working directory. This breaks the
ability to use a 'clouds.yaml' file in the local directory when
provisioning on OpenStack clouds.

Restore this by checking for the presence of such a file early in the
provisioning process and setting an environment variable to point to
this file if found. This isn't the nicest solution - it would probably
be better to load the file as an asset and copy it to the Terraform
working directory - but it should be effective.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>