Download `yq` in upi installer containers
Various CI steps use the upi-installer container for it's access to the
aws cli tools among other things.  However, most of those steps also
curl `yq` directly from GitHub.  We can save ourselves some headaches
when GitHub is down by just embedding the binary in the image already.

Note, this also removes the python yq which I am not sure if it's used,
but I can't find a place where it is.