aws: Fix network interface not found error
Right now, the control-plane nodes and their network interfaces
are being created in parallel and sometimes the control plane nodes
try to find their network interface before it's created, causing
an error in the creation of the nodes. Adding a depends on block
to instruct terraform to create the network interfaces before
trying to create the control plane nodes.