AGENT-228: Systemd service to set hostname defined in agent-config
Users may specify a hostname for a host by defining the hostname
and the host's macAddress in agent-config.yaml.

The agent Ignition then writes each macAddress to hostname mapping
to a file at /etc/assisted-installer/hostnames/.

The file names in that path are MAC addresses. The desired hostname
is the file's content.

A new set-hostname.service runs during boot and reads the macAddress
to hostname mappings and if it finds one of the host's macAddress
is in the mapping, it then updates the host's hostname using the
mapped value.