AGENT-230: Enable multipathd in agent ISO
Enable the multipathd service with a default config file in the agent
ISO. This is equivalent to specifying the `rd.multipath=default` flag on
the kernel command line, but without the need to mess with the kernel
command line. Since the ISO itself will be attached as virtual or actual
media, there is no need for anything to run during the initramfs.

This allows us the agent to discover multipath Fibre Channel volumes if
they exist.