installconfig: Move fatal error to warn for strict marshal
The hive team uses a lot of fields in the install config that might
not exist due to older versions of the installer missing a few fields
that may have been added in the newer versions, which causes the
installer to reject them all but is needed for proper functioning.

Changing the code a little to now provide a warning message to the
user in case of any unknown fields and then try to unmarshal it
without the unknown fields check.