Delete Pod by name rather than ID
I've occasionally seen a failure cause us to end up with an
assisted-service pod created in podman, but the service fails. For some
reason the pod does not get deleted (possibly the deletion command
itself fails), and subsequent attempts to restart
the service are unable to create a new pod with the same name, but they
do delete the pod ID file and replace it with an empty one.

Perhaps it will be more robust to delete by name, since this will mean
that subsequent retries could possibly succeed if they do run the
ExecStopPost script.