gcp: Fix panic when unknown region is provided.
If an unknown region is provided, installer queries GCP for zones
which returns empty. The installer assumes if there are no errors,
the array is populated and uses the first entry that does not
get populated. Adding another safe check to see if the zone length
is not zero and then proceed.