AGENT-263: Get images from release payload
Get the images used to start all services from the release payload which may be a
mirror. Since we do not build a version of the UI image, that has been removed. Also
for the assisted-service-db the postgresql image is no longer used, instead the
assisted-service image is used to start the database.