Bug 2106820: Convert all MAC addresses to lowercase
If MAC addresses are provided with uppercase hex digits in the
agent-config, convert them to lowercase before storing in the image. The
client program that searches for hosts just does a direct
(case-sensitive) comparison with the values obtained from inventory. The
service that sets the hostname also does a case-sensitive comparison
with the output of `ip address`.