Extract release's image-references file for MCO
Previously all the MCO needed was the machine-os-content image, which
became osImageURL, but now that we have new format images, extension
containers, and probably a pair (one new format image, one extension
container) per RHCOS major version (rhel-coreos-8, rhel-coreos-9, etc)
we need to consume more than one image out of the payload.

As a result, the MCO needs the image-references file so that it can read all of the
OS images contained within and select the one it prefers.

This dumps the image-references file out of the payload so the MCO can
consume it, and also adds an argument to the `machine-config-operator`
call to consume it.