AGENT-286: Use local image instead of podman pause container
The 'podman create pod' requires a pause container which is fetched from
registry.access.redhat.com/ubi8/pause:latest. This won't work in a disconnected
environment. This change overwrite the infra image which is used for the
'pause' command and uses the assisted-service image instead which is in the
payload. Since 'pause' just pends, the command is also changed to do just that.