logging: Decorate logs with fields
The requirement is to add extra fields to every log field for ARO
to identify the cluster using these tags. Adding an option to
provide a file "log-config.yaml" file in the install directory
which would have the fields and a log level value.

Every log that has the level below or equal to the log level
provided in the file will be decorated with the key-value field tags
in the file.